+++ src/DpressServices.php
@@ -27,6 +27,7 @@ use Dynart\Micro\Entities\PdoBuilder;
 use Dynart\Micro\Entities\QueryBuilder;
 use Dynart\Micro\Entities\QueryBuilder\MariaQueryBuilder;
 use Dynart\Micro\Entities\QueryExecutor;
+use Dynart\Dpress\Cli\BundleCommands;
 use Dynart\Dpress\Cli\DoctorCommands;
 use Dynart\Dpress\Cli\SchemaCommands;
 use Dynart\Dpress\Cli\SystemCommands;
@@ -115,6 +116,7 @@ use Dynart\Dpress\Content\Sitemap;
 use Dynart\Dpress\Content\PageContext;
 use Dynart\Dpress\Service\RoleService;
 use Dynart\Dpress\Service\Doctor;
+use Dynart\Dpress\Service\SiteBundle;
 use Dynart\Dpress\Service\SchemaService;
 use Dynart\Dpress\Service\UserService;
 
@@ -208,6 +210,7 @@ class DpressServices {
         Micro::add(RateLimiter::class);
         Micro::add(SchemaService::class);
         Micro::add(Doctor::class);
+        Micro::add(SiteBundle::class);
         Micro::add(RoleService::class);
         Micro::add(UserService::class);
         Micro::add(AuthService::class);
@@ -253,6 +256,7 @@ class DpressServices {
         Micro::add(MarkdownBlock::class);
         Micro::add(SchemaCommands::class);
         Micro::add(DoctorCommands::class);
+        Micro::add(BundleCommands::class);
         Micro::add(SystemCommands::class);
         Micro::add(UserCommands::class);
         Micro::add(MailCommands::class);